Extract performance metrics into SRP microcrate
