+++ Cargo.toml
@@ -4,6 +4,7 @@ members = [
   "crates/bitnet-common",
   "crates/bitnet-math",
   "crates/bitnet-warn-once",
+  "crates/bitnet-performance-metrics-core", # SRP: shared performance metric + computation types
   "crates/bitnet-runtime-context",
   "crates/bitnet-runtime-context-core",
   "crates/bitnet-bdd-grid",
@@ -61,6 +62,7 @@ members = [
   "crates/bitnet-startup-contract-guard",
   "crates/bitnet-test-support",
   "crates/bitnet-validation",    # SRP: LayerNorm/projection validation rules (shared)
+  "crates/bitnet-performance-metrics-core",
   "crates/bitnet-gpu-hal",       # GPU hardware abstraction layer
   "crates/bitnet-opencl",        # OpenCL backend for Intel Arc GPU inference
   "crates/bitnet-metal",         # MSL compute kernels for Apple Silicon
@@ -85,6 +87,7 @@ default-members = [
   "crates/bitnet-common",
   "crates/bitnet-math",
   "crates/bitnet-warn-once",
+  "crates/bitnet-performance-metrics-core", # SRP: shared performance metric + computation types
   "crates/bitnet-quantization",
   "crates/bitnet-rope",
   "crates/bitnet-transformer",
@@ -114,6 +117,7 @@ default-members = [
   "crates/bitnet-engine-core",
   "crates/bitnet-gguf",
   "crates/bitnet-validation",    # SRP: LayerNorm/projection validation rules (shared)
+  "crates/bitnet-performance-metrics-core",
 ]
 
 [package]
@@ -160,6 +164,7 @@ version = "0.2.1-dev"
 bitnet-common = { path = "crates/bitnet-common", version = "0.2.1-dev" }
 bitnet-math = { path = "crates/bitnet-math", version = "0.2.1-dev" }
 bitnet-warn-once = { path = "crates/bitnet-warn-once", version = "0.2.1-dev" }
+bitnet-performance-metrics-core = { path = "crates/bitnet-performance-metrics-core", version = "0.2.1-dev" }
 bitnet-models = { path = "crates/bitnet-models", version = "0.2.1-dev" }
 bitnet-quantization = { path = "crates/bitnet-quantization", version = "0.2.1-dev", default-features = false }
 bitnet_ggml_ffi = { package = "bitnet-ggml-ffi", path = "crates/bitnet-ggml-ffi", version = "0.2.1-dev", optional = true }
@@ -321,6 +326,7 @@ required-features = ["cpu"]
 [workspace.dependencies]
 bitnet-math = { path = "crates/bitnet-math", version = "0.2.1-dev" }
 bitnet-warn-once = { path = "crates/bitnet-warn-once", version = "0.2.1-dev" }
+bitnet-performance-metrics-core = { path = "crates/bitnet-performance-metrics-core", version = "0.2.1-dev" }
 # Core dependencies
 anyhow = "1.0.100"
 log = "0.4.28"
