+++ crates/bitnet-common/Cargo.toml
@@ -18,6 +18,7 @@ include = [
 [dependencies]
 bitnet-math.workspace = true
 bitnet-warn-once.workspace = true
+bitnet-performance-metrics-core.workspace = true
 anyhow.workspace = true
 thiserror.workspace = true
 serde.workspace = true
