+++ crates/bitnet-common/src/strict_mode.rs
@@ -4,6 +4,7 @@
 //! and ensure real quantized computation is used throughout the inference pipeline.
 
 use crate::{BitNetError, Result};
+pub use bitnet_performance_metrics_core::{ComputationType, PerformanceMetrics};
 use std::env;
 use std::sync::OnceLock;
 
@@ -280,30 +281,6 @@ pub struct MissingKernelScenario {
     pub quantization_type: crate::QuantizationType,
     pub device: crate::Device,
     pub fallback_available: bool,
-}
-
-/// Performance metrics for validation
-#[derive(Debug, Clone, Default, serde::Serialize, serde::Deserialize)]
-pub struct PerformanceMetrics {
-    #[serde(default)]
-    pub tokens_per_second: f64,
-    #[serde(default)]
-    pub latency_ms: f64,
-    #[serde(default)]
-    pub memory_usage_mb: f64,
-    #[serde(default)]
-    pub computation_type: ComputationType,
-    #[serde(default)]
-    pub gpu_utilization: Option<f64>,
-}
-
-/// Computation type classification
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Default, serde::Serialize, serde::Deserialize)]
-#[serde(rename_all = "snake_case")]
-pub enum ComputationType {
-    #[default]
-    Real,
-    Mock,
 }
 
 #[cfg(test)]
