+++ crates/bitnet-performance-metrics-core/Cargo.toml
@@ -1,0 +1,17 @@
+[package]
+name = "bitnet-performance-metrics-core"
+version.workspace = true
+edition.workspace = true
+license.workspace = true
+repository.workspace = true
+homepage.workspace = true
+authors.workspace = true
+keywords.workspace = true
+categories.workspace = true
+description = "SRP microcrate for shared performance metric types and computation classification"
+
+[dependencies]
+serde.workspace = true
+
+[dev-dependencies]
+serde_json.workspace = true
